Don't always replace the custom color scheme.

+++ install.sh
@@ -72,8 +72,10 @@ sed -i 's~\(ZSH_THEME="\)[^"]*\(".*\)~\1powerlevel10k/powerlevel10k\2~' ~/.zshrc
 curl -fsSL -o ~/.termux/font.ttf 'https://github.com/romkatv/dotfiles-public/raw/master/.local/share/fonts/NerdFonts/MesloLGS%20NF%20Regular.ttf'
 
 # Set 'Tango' as the default color scheme for the shell.
+if [ ! -e ~/.termux/colors.properties ]; then
 cp -fr "$HOME/.oh-my-zsh/custom/misc/LitMux/.termux/colors.properties" ~/.termux/colors.properties
 termux-reload-settings
+fi
 
 # Replace the default welcome text with a customized one.
 cp -fr "$HOME/.oh-my-zsh/custom/misc/LitMux/motd-lit" "$PREFIX/etc/motd"
